+++ src/main/java/dgu/newsee/domain/news/controller/UserNewsController.java
@@ -2,10 +2,8 @@ package dgu.newsee.domain.news.controller;
 
 import dgu.newsee.domain.news.dto.NewsDetailDTO;
 import dgu.newsee.domain.news.dto.NewsDTO;
-import dgu.newsee.domain.news.dto.SavedNewsDTO;
 import dgu.newsee.domain.news.service.NewsService;
-import dgu.newsee.domain.user.entity.Level;
-import dgu.newsee.domain.user.entity.User;
+import dgu.newsee.domain.news.util.SecurityUtil;
 import dgu.newsee.domain.user.repository.UserRepository;
 import dgu.newsee.global.exception.NewsException;
 import dgu.newsee.global.payload.ApiResponse;
@@ -28,38 +26,28 @@ public class UserNewsController {
     @GetMapping
     @Operation(summary = "전체 뉴스 조회 (비회원은 중 레벨)")
     public ApiResponse<List<NewsDTO>> getAllNews(Authentication authentication) {
-        Long userId = extractUserId(authentication);
+        Long userId = SecurityUtil.extractUserId(authentication);
         return newsService.getAllNews(userId, null);
     }
 
     @GetMapping("/{newsId}")
     @Operation(summary = "뉴스 상세 조회 (비회원은 중 레벨)")
     public ApiResponse<NewsDetailDTO> getNewsDetail(@PathVariable Long newsId, Authentication authentication) {
-        Long userId = extractUserId(authentication);
+        Long userId = SecurityUtil.extractUserId(authentication);
         return newsService.getNewsDetail(newsId, userId, null);
     }
 
     @GetMapping("/search")
     @Operation(summary = "뉴스 검색 (비회원은 중 레벨)")
     public ApiResponse<List<NewsDTO>> searchNews(@RequestParam String keyword, Authentication authentication) {
-        Long userId = extractUserId(authentication);
+        Long userId = SecurityUtil.extractUserId(authentication);
         return newsService.searchNews(keyword, userId, null);
-    }
-
-    @GetMapping("/user/saved-news")
-    @Operation(summary = "북마크한 뉴스 목록 조회 (회원만 가능)")
-    public ApiResponse<List<SavedNewsDTO>> getSavedNews(Authentication authentication) {
-        Long userId = extractUserId(authentication);
-        if (userId == null) {
-            throw new NewsException(ResponseCode.USER_UNAUTHORIZED);
-        }
-        return newsService.getSavedNews(userId, null);
     }
 
     @PostMapping("/{newsId}")
     @Operation(summary = "뉴스 북마크 저장 (회원만 가능)")
     public ApiResponse<?> saveNews(@PathVariable Long newsId, Authentication authentication) {
-        Long userId = extractUserId(authentication);
+        Long userId = SecurityUtil.extractUserId(authentication);
 
         if (userId == null) {
             throw new NewsException(ResponseCode.USER_UNAUTHORIZED);
@@ -74,7 +62,7 @@ public class UserNewsController {
             @PathVariable Long savedNewsId,
             Authentication authentication
     ) {
-        Long userId = extractUserId(authentication);
+        Long userId = SecurityUtil.extractUserId(authentication);
 
         if (userId == null) {
             throw new NewsException(ResponseCode.USER_UNAUTHORIZED);
@@ -83,10 +71,4 @@ public class UserNewsController {
         return newsService.deleteNewsBookmark(userId, savedNewsId);
     }
 
-    private Long extractUserId(Authentication authentication) {
-        if (authentication == null || authentication.getName() == null) {
-            return null;
-        }
-        return Long.parseLong(authentication.getName());
-    }
 }
